Refactor code structure for improved readability and maintainability

+++ astraweave-ai/tests/perception_tests.rs
@@ -55,10 +55,10 @@ fn create_complex_snapshot(time: f32) -> WorldSnapshot {
         enemies.push(EnemyState {
             id: 200 + i as u32,
             pos: IVec2 {
-                x: 20 + (i as i32) * 2,
-                y: 15 + (i as i32 % 3),
+                x: 20 + i * 2,
+                y: 15 + (i % 3),
             },
-            hp: 50 + (i as i32) * 5,
+            hp: 50 + i * 5,
             cover: if i % 2 == 0 { "low" } else { "high" }.to_string(),
             last_seen: (i as f32) * 0.5,
         });
@@ -69,8 +69,8 @@ fn create_complex_snapshot(time: f32) -> WorldSnapshot {
         pois.push(Poi {
             k: if i % 2 == 0 { "medkit" } else { "ammo" }.to_string(),
             pos: IVec2 {
-                x: 5 + (i as i32) * 3,
-                y: 5 + (i as i32) * 2,
+                x: 5 + i * 3,
+                y: 5 + i * 2,
             },
         });
     }
@@ -78,8 +78,8 @@ fn create_complex_snapshot(time: f32) -> WorldSnapshot {
     let mut obstacles = vec![];
     for i in 0..30 {
         obstacles.push(IVec2 {
-            x: 10 + (i as i32) % 10,
-            y: 10 + (i as i32) / 10,
+            x: 10 + (i % 10),
+            y: 10 + (i / 10),
         });
     }
